:recycle: [Refactor] mock.ts 삭제 type 추가
